requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiosqlite==0.16.1
async-timeout==3.0.1
attrs==21.2.0
better-exceptions==0.3.3
cachetools==4.2.2
chardet==4.0.0
choicelib==0.1.4
click==7.1.2
colorama==0.4.4
huepy==1.2.1
idna==3.2
iso8601==0.1.16
Jinja2==2.11.3
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
pydantic==1.8.2
Pygments==2.9.0
pypika-tortoise==0.1.1
pytz==2021.1
tortoise-orm==0.17.6
typer==0.3.2
typing-extensions==3.10.0.0
vbml==1.1
vkquick @ https://github.com/deknowny/vkquick/archive/master.zip
watchgod==0.6
win32-setctime==1.0.3
yarl==1.6.3
